requirements.txt
discord.py[voice]
aiohttp
asyncpg
apscheduler
sqlalchemy
pip
